[testing] Move test vsync waiters to their own TUs
This move makes it easier to add more vsync waiters.
