+++ shell/common/shell_test.cc
@@ -289,46 +289,6 @@ void ShellTest::AddNativeCallback(std::string name,
   native_resolver_->AddNativeCallback(std::move(name), callback);
 }
 
-void ShellTestVsyncClock::SimulateVSync() {
-  std::scoped_lock lock(mutex_);
-  if (vsync_issued_ >= vsync_promised_.size()) {
-    vsync_promised_.emplace_back();
-  }
-  FML_CHECK(vsync_issued_ < vsync_promised_.size());
-  vsync_promised_[vsync_issued_].set_value(vsync_issued_);
-  vsync_issued_ += 1;
-}
-
-std::future<int> ShellTestVsyncClock::NextVSync() {
-  std::scoped_lock lock(mutex_);
-  vsync_promised_.emplace_back();
-  return vsync_promised_.back().get_future();
-}
-
-void ShellTestVsyncWaiter::AwaitVSync() {
-  FML_DCHECK(task_runners_.GetUITaskRunner()->RunsTasksOnCurrentThread());
-  auto vsync_future = clock_.NextVSync();
-
-  auto async_wait = std::async([&vsync_future, this]() {
-    vsync_future.wait();
-
-    // Post the `FireCallback` to the Platform thread so earlier Platform tasks
-    // (specifically, the `VSyncFlush` call) will be finished before
-    // `FireCallback` is executed. This is only needed for our unit tests.
-    //
-    // Without this, the repeated VSYNC signals in `VSyncFlush` may start both
-    // the current frame in the UI thread and the next frame in the secondary
-    // callback (both of them are waiting for VSYNCs). That breaks the unit
-    // test's assumption that each frame's VSYNC must be issued by different
-    // `VSyncFlush` call (which resets the `will_draw_new_frame` bit).
-    //
-    // For example, HandlesActualIphoneXsInputEvents will fail without this.
-    task_runners_.GetPlatformTaskRunner()->PostTask([this]() {
-      FireCallback(fml::TimePoint::Now(), fml::TimePoint::Now());
-    });
-  });
-}
-
 ShellTestPlatformView::ShellTestPlatformView(PlatformView::Delegate& delegate,
                                              TaskRunners task_runners,
                                              bool simulate_vsync)
