+++ shell/common/shell_test.h
@@ -13,6 +13,7 @@
 #include "flutter/lib/ui/window/platform_message.h"
 #include "flutter/shell/common/run_configuration.h"
 #include "flutter/shell/common/shell.h"
+#include "flutter/shell/common/test_vsync_waiters.h"
 #include "flutter/shell/common/thread_host.h"
 #include "flutter/shell/gpu/gpu_surface_gl_delegate.h"
 #include "flutter/testing/test_dart_native_resolver.h"
@@ -86,32 +87,6 @@ class ShellTest : public ThreadTest {
   fml::UniqueFD assets_dir_;
 
   FML_DISALLOW_COPY_AND_ASSIGN(ShellTest);
-};
-
-class ShellTestVsyncClock {
- public:
-  /// Simulate that a vsync signal is triggered.
-  void SimulateVSync();
-
-  /// A future that will return the index the next vsync signal.
-  std::future<int> NextVSync();
-
- private:
-  std::mutex mutex_;
-  std::vector<std::promise<int>> vsync_promised_;
-  size_t vsync_issued_ = 0;
-};
-
-class ShellTestVsyncWaiter : public VsyncWaiter {
- public:
-  ShellTestVsyncWaiter(TaskRunners task_runners, ShellTestVsyncClock& clock)
-      : VsyncWaiter(std::move(task_runners)), clock_(clock) {}
-
- protected:
-  void AwaitVSync() override;
-
- private:
-  ShellTestVsyncClock& clock_;
 };
 
 class ShellTestPlatformView : public PlatformView, public GPUSurfaceGLDelegate {
